Trait kdevice::block::BlockBus

pub trait BlockBus {
    fn probe(&mut self) -> Result<()>;

    fn read(&mut self, bno: usize, buf: &mut [u8]) -> Result<()>;

    fn write(&mut self, bno: usize, buf: &[u8]) -> Result<()>;
}

Block device trait that you need to implement for each slot.

For each instance, all three methods are guaranteed to becalled exlusively.

Required methods

fn probe(&mut self) -> Result<()>

Return Ok if the bus is active.

fn read(&mut self, bno: usize, buf: &mut [u8]) -> Result<()>

Read a block from the block device.

fn write(&mut self, bno: usize, buf: &[u8]) -> Result<()>

Write a block to the block device.