Trait kdevice::cache::Cache

pub trait Cache where
    Self: Sized, {
    type Key: Hash + Eq + Default + Copy + Debug;
    type Value: Debug;
    type ReadFut: Future<Output = Result<()>> + 'a;
    type WriteFut: Future<Output = Result<()>> + 'a;
    type GetFut: Future<Output = Result<Option<CEntry<'a, Self>>>> + 'a;
    fn cache_self<'a>(
        &'a self
    ) -> &'a Spinlock<CacheData<Self::Key, Self::Value>>;

    fn disk_read<'a>(
        &'a self, 
        key: &'a Self::Key, 
        val: &'a mut Self::Value
    ) -> Self::ReadFut;

    fn disk_write<'a>(
        &'a self, 
        key: &'a Self::Key, 
        val: &'a Self::Value
    ) -> Self::WriteFut;

    fn cache_get<'a>(&'a self, key: Self::Key, flush: bool) -> Self::GetFut;

    fn cache_invalidate<'a>(&'a self);

    fn new_cache(
        n: usize, 
        default_val: impl Fn() -> Result<Self::Value>
    ) -> Result<Spinlock<CacheData<Self::Key, Self::Value>>> { ... }
}

Cache trait.

Associated Types

type Key: Hash + Eq + Default + Copy + Debug

Key type.

type Value: Debug

Value type.

type ReadFut: Future<Output = Result<()>> + 'a

Future returned by Self::disk_read.

type WriteFut: Future<Output = Result<()>> + 'a

Future returned by Self::disk_write.

type GetFut: Future<Output = Result<Option<CEntry<'a, Self>>>> + 'a

Future returned by Self::cache_get.

Required methods

fn cache_self<'a>(&'a self) -> &'a Spinlock<CacheData<Self::Key, Self::Value>>

Function used to retrieve the cache.

fn disk_read<'a>(
&'a self,
key: &'a Self::Key,
val: &'a mut Self::Value
) -> Self::ReadFut

Async function to fetch data from disk, should be implemented by yourself.

fn disk_write<'a>(
&'a self,
key: &'a Self::Key,
val: &'a Self::Value
) -> Self::WriteFut

Async function to flush data to disk, should be implemented by yourself.

fn cache_get<'a>(&'a self, key: Self::Key, flush: bool) -> Self::GetFut

Get a cache item by key. Won’t lock it. If key has already been in memory, this operation always succeed. If cannot reserve cache entry, returns None. Donot implement it. Instead, injected to the trait implementation by super::cache_impl.

fn cache_invalidate<'a>(&'a self)

Mark all dirty entry as invalid without synchronizing with disk. May cause inconsistency! Donot implement it. Instead, injected to the trait implementation by super::cache_impl.

Provided methods

fn new_cache(
n: usize,
default_val: impl Fn() -> Result<Self::Value>
) -> Result<Spinlock<CacheData<Self::Key, Self::Value>>>

Create a new cache with at most n in-memory cache entries.

default_val is used to generate the initial value. It should return a Result since oom may occur during generation, e.g. Box<[u8; 512]>.

Donot override this implementation.