Struct kdevice::fat::FAT

pub struct FAT { /* fields omitted */ }

TFAT32

Implementations

impl FAT

pub async fn new(ninodes: usize, nbuf: usize, disk: &Arc<Chan>) -> Result<Self>

Initialize a new FAT from a virtual disk.

ninodes is the maximum active inode, nbuf is the number of cached buffer.

pub fn to_inode<'a>(&'a self, path: u64) -> ManuallyDrop<CEntry<'a, Self>>

Convert qid to inode pointer.

pub fn to_path<'a>(&'a self, ip: ManuallyDrop<CEntry<'a, Self>>) -> u64

Convert inode pointer to unique path id.

pub async fn iget<'a>(
&'a self,
key: InodeKey
) -> Option<ManuallyDrop<CEntry<'a, Self>>>

Open and increase the reference of a file.

pub fn iput<'a>(&'a self, ip: ManuallyDrop<CEntry<'a, Self>>)

Close and decrease the reference counter of a file.

pub async fn fatrw<'a>(
&'a self,
cno: u32,
f: impl FnOnce(u32) -> u32
) -> Result<u32>

Read or modify a FAT entry. Return the old value.

If f is None, read the fat value. Otherwise apply f on the value of FAT entry. If failed the FAT entry won’t be updated in both bio cache and log.

Return the old entry value.

pub async fn calloc<'a>(&'a self, n: usize) -> Result<u32>

Allocate a chain of cluster of len n.

Return validated head cluster number.

pub async fn removei<'a>(&'a self, ip: &'a CGuard<'a, Self>) -> Result<()>

This should only be called when

No other transactions.

The size of file is within one cluster.

A plain file or empty directory.

Modify at most 2 blocks.

pub async fn stati<'a>(&'a self, ip: &'a CGuard<'a, Self>) -> Result<Dirent>

Get the meta information of this file.

pub async fn extend<'a>(&'a self, ip: &mut CGuard<'a, Self>) -> Result<()>

Extend the addr pointers.

pub async fn resize<'a>(
&'a self,
ip: &mut CGuard<'a, Self>,
f: impl FnOnce(u32) -> u32
) -> Result<u32>

Modify the size of this file and update its cluster pointers.

Return the new size. For directory, the old size is always a multiple of byte-per-cluster.

pub async fn readi<'a>(
&'a self,
ip: &mut CGuard<'a, Self>,
buf: &mut [u8],
off: usize
) -> Result<usize>

Read data from inode.

pub async fn writei<'a>(
&'a self,
ip: &mut CGuard<'a, Self>,
buf: &[u8],
off: usize
) -> Result<usize>

Write data to inode.

pub async fn dirlookup<'a>(
&'a self,
dp: &CGuard<'a, Self>,
name: &[u8]
) -> Result<Option<InodeKey>>

Look up a file in directory. Return a referenced inode.

pub async fn dirlink<'a>(
&'a self,
dp: &mut CGuard<'a, Self>,
name: &[u8],
cno: Option<u32>,
dir: bool
) -> Result<Option<InodeKey>>

Create a directory entry with specific inum.

Caller should guarantee then name doesn’t exist in this dir. Return inode if the link succeed. Return None if entry with the same name already exist.

pub async fn dirempty<'a>(&'a self, dp: &'a CGuard<'a, Self>) -> Result<bool>

Check if directory is empty.