Struct kdevice::log::Log

pub struct Log { /* fields omitted */ }

Generic log layer with buffer cache.

Implementations

impl Log

pub async fn new(nbuf: usize, start: usize, disk: Arc<Chan>) -> Result<Self>

Create a buffer cache with maximum nbuf in-memory buffers.

Close the disk on error.

pub async fn close(self)

Close the underlying disk file. Async destructor.

pub async fn begin_opx(&self, reserved: usize) -> Result<()>

Called at the start of each transaction.

pub async fn begin_exop(&self) -> Result<()>

Start an exclusive operation.

At any time, only one exclusive operation or some normal operations are allowed to be active.

pub async fn end_exop<'a>(&'a self, bad: bool, wait: bool) -> OpGuard<'a>

End an exlusive operation.

pub async fn begin_op(&self) -> Result<()>

Called at the start of each transaction.

pub async fn end_op<'a>(&'a self, bad: bool, wait: bool) -> OpGuard<'a>

Called at the end of each operation.

Try to commit if this was the last outstanding operation. Otherwise wait until transaction completes.

If mark as bad, every operation in this transaction won’t execute.

If wait is true, new transactions won’t enter begin_op until all waiters of current transaction have dropped their guards. This is useful when you want to block new operations and clean something when the current transaction is bad.

Return whether the whole transaction is actually committed or not.

pub async fn trace<'a>(&'a self, buf: CacheGuard<'a, Self>) -> usize

Keep track of a buffer in log.

Return the number of newly traced blocks (0 or 1).

NOTE:

Must use cache_get without flush to avoid emitting dirty blocks, which is handled by log.

Must use this function instead of unlock.